requirements.txt
Flask==2.3.2
pdf2image==1.16.3
pdfplumber==0.9.0
Pillow==9.5.0
PyPDF2==3.0.1
pytesseract==0.3.10
tabula-py==2.7.0